test the upgrade prompt for legacy product search block
